fix: メインの内容をホームページだけに表示することが可能
